Issue #2540: fix spelling errors in it packages
